requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
groq>=0.5.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
streamlit>=1.32.0
pydantic>=2.6.0
python-dotenv>=1.0.0
mcp>=1.0.0
httpx>=0.27.0
